+++ src/types.ts
@@ -56,7 +56,7 @@ function validBoolean(value: unknown, fallback: boolean): boolean {
 
 function validBinding(value: unknown): string | null {
   return typeof value === "string" &&
-    /^(code|key):.+$/.test(value) &&
+    /^(code|key|keyCode|keyIdentifier):.+$/.test(value) &&
     value !== "code:Unidentified" &&
     value !== "key:Unidentified"
     ? value